Use a netlify lambda for the API
