--- Procfile
@@ -1,1 +1,0 @@
-web: node -r esm index.js
