--- index.js
@@ -1,17 +1,0 @@
-import resolvers from './resolvers';
-import typeDefs from './schema';
-import {ApolloServer} from 'apollo-server';
-
-const server = new ApolloServer({
-  typeDefs,
-  resolvers,
-  introspection: true,
-  playground: true,
-  engine: {
-    apiKey: process.env.ENGINE_API_KEY
-  }
-});
-
-server.listen({port: process.env.PORT}).then(({url}) => {
-  console.log(`🚀  Server ready at ${url}`);
-});
