+++ static/_redirects
@@ -1,0 +1,1 @@
+/ /.netlify/functions/graphql
